# WebContent/doc/user/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": ""
  },
  "host": "localhost:8080",
  "basePath": "/RnsSystem/rest",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/connections": {
      "get": {
        "tags": [
          "connections"
        ],
        "summary": "Get connections",
        "description": "Get connections of rns, restricted to admin and returned in portion",
        "operationId": "getConnections",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "name": "admin",
            "in": "query",
            "description": "Specify if the client requesting the resource is an admin",
            "required": false,
            "type": "boolean"
          },
          {
            "name": "page",
            "in": "query",
            "description": "Which page of the resource must be returned",
            "required": false,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Connections"
            }
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/places": {
      "get": {
        "tags": [
          "places"
        ],
        "summary": "Get places",
        "description": "Get places of rns, returned in portion",
        "operationId": "getPlaces",
        "consumes": [
          "text/plain"
        ],
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Which page of the resource must be returned",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "idSuffix",
            "in": "query",
            "description": "Retrieve only places for which the id have this suffix",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Places"
            }
          }
        }
      }
    },
    "/places/gates": {
      "get": {
        "tags": [
          "places"
        ],
        "summary": "Get gates",
        "description": "Get gates in the system, restricted to admin",
        "operationId": "getGates",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "name": "admin",
            "in": "query",
            "description": "Specify if the client requesting the resource is an admin",
            "required": false,
            "type": "boolean",
            "default": "false"
          },
          {
            "name": "gateType",
            "in": "query",
            "description": "Get only gates of the specified type",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Gates"
            }
          },
          "400": {
            "description": "Bad request"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/places/parkingAreas": {
      "get": {
        "tags": [
          "places"
        ],
        "summary": "Get parking areas",
        "description": "Get parking areas in the system, restricted to admin",
        "operationId": "getParkingAreas",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "name": "admin",
            "in": "query",
            "description": "Specify if the client requesting the resource is an admin",
            "required": false,
            "type": "boolean",
            "default": "false"
          },
          {
            "name": "service",
            "in": "query",
            "description": "Get only parking areas that have the specified list of services",
            "required": false,
            "type": "array",
            "items": {
              "type": "string"
            },
            "collectionFormat": "csv"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ParkingAreas"
            }
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/places/roadSegments": {
      "get": {
        "tags": [
          "places"
        ],
        "summary": "Get road segments",
        "description": "Get road segments in the system, restricted to admin",
        "operationId": "getRoadSegments",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "name": "admin",
            "in": "query",
            "description": "Specify if the client requesting the resource is an admin",
            "required": false,
            "type": "boolean",
            "default": "false"
          },
          {
            "name": "roadName",
            "in": "query",
            "description": "Get road segment only from the specified road name",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/RoadSegments"
            }
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/places/{id}": {
      "get": {
        "tags": [
          "places"
        ],
        "summary": "Get place",
        "description": "Get a single place by its id, restricted to admin",
        "operationId": "getPlace",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "name": "admin",
            "in": "query",
            "description": "Specify if the client requesting the resource is an admin",
            "required": false,
            "type": "boolean",
            "default": "false"
          },
          {
            "name": "id",
            "in": "path",
            "description": "Id of the place to get",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/PlaceType"
            }
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "definitions": {
    "Connections": {
      "type": "object",
      "properties": {
        "totalPages": {
          "type": "integer",
          "format": "int64"
        },
        "page": {
          "type": "integer",
          "format": "int64"
        },
        "connection": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Connection"
          }
        }
      },
      "xml": {
        "name": "connections"
      }
    },
    "ParkingAreas": {
      "type": "object",
      "properties": {
        "totalPages": {
          "type": "integer",
          "format": "int64"
        },
        "page": {
          "type": "integer",
          "format": "int64"
        },
        "parkingArea": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ParkingArea"
          }
        }
      },
      "xml": {
        "name": "parkingAreas"
      }
    },
    "PlaceType": {
      "type": "object",
      "properties": {
        "connection": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "id": {
          "type": "string"
        },
        "self": {
          "type": "string"
        },
        "capacity": {
          "type": "integer",
          "format": "int64"
        },
        "vehicles": {
          "type": "string"
        }
      }
    },
...
    "Connection": {
      "type": "object",
      "required": [
        "from",
        "to"
      ],
      "properties": {
        "from": {
          "type": "string"
        },
        "to": {
          "type": "string"
        }
      },
      "xml": {
        "name": "connection"
      }
    },
    "RoadSegment": {
      "type": "object",
      "properties": {
        "connection": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "id": {
          "type": "string"
        },
        "self": {
          "type": "string"
        },
        "capacity": {
          "type": "integer",
          "format": "int64"
        },
        "vehicles": {
          "type": "string"
        },
        "roadName": {
          "type": "string"
        },
        "name": {
          "type": "string"
        }
      }
    },
    "RoadSegments": {
      "type": "object",
      "properties": {
        "totalPages": {
          "type": "integer",
          "format": "int64"
        },
        "page": {
          "type": "integer",
          "format": "int64"
        },
        "roadSegment": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/RoadSegment"
          }
        }
      },
      "xml": {
        "name": "roadSegments"
      }
    },
...
    "Gate": {
      "type": "object",
      "properties": {
        "connection": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "id": {
          "type": "string"
        },
        "self": {
          "type": "string"
        },
        "capacity": {
          "type": "integer",
          "format": "int64"
        },
        "vehicles": {
          "type": "string"
        },
        "type": {
          "type": "string",
          "enum": [
            "IN",
            "OUT",
            "INOUT"
          ]
        }
      }
    },
...
    "Places": {
      "type": "object",
      "required": [
        "gates",
        "parkingAreas",
        "roadSegments"
      ],
      "properties": {
        "totalPages": {
          "type": "integer",
          "format": "int64"
        },
        "page": {
          "type": "integer",
          "format": "int64"
        },
        "roadSegments": {
          "type": "string"
        },
        "parkingAreas": {
          "type": "string"
        },
        "gates": {
          "type": "string"
        },
        "place": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/PlaceType"
          }
        }
      },
      "xml": {
        "name": "places"
      }
    },
...
    "ParkingArea": {
      "type": "object",
      "required": [
        "services"
      ],
      "properties": {
        "connection": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "id": {
          "type": "string"
        },
        "self": {
          "type": "string"
        },
        "capacity": {
          "type": "integer",
          "format": "int64"
        },
        "vehicles": {
          "type": "string"
        },
        "services": {
          "$ref": "#/definitions/ServicesType"
        }
      }
    },
    "ServicesType": {
      "type": "object",
      "properties": {
        "service": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
...
    "Gates": {
      "type": "object",
      "properties": {
        "totalPages": {
          "type": "integer",
          "format": "int64"
        },
        "page": {
          "type": "integer",
          "format": "int64"
        },
        "gate": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Gate"
          }
        }
      },
      "xml": {
        "name": "gates"
      }
    }
  }
}